requirements.txt
Django>=2.2.7,<2.3.0
djangorestframework>=3.10.3,<3.11.0
psycopg2>=2.8.4,<2.9.0
flake8>=3.7.9,<3.8.0
request>=2019.4.13,<2019.5.0
drf-yasg>=1.17.0,<1.18
django-cors-headers>=3.2.0,<3.3.0

